LA-52: (Client) Update new UI code to point to renamed Accounts API source file.
