+++ client/src/components/GeneralJournalEntry.js
@@ -3,7 +3,7 @@ import { Nav, NavItem } from 'react-bootstrap';
 import './GeneralJournal.css';
 import './CommonChart.css';
 import GeneralJournalAPI from '../api/GeneralJournal.js';
-import AccountsAPI from '../api/Accounts.js';
+import AccountsAPI from '../api/AccountsApi.js';
 
 import JournalEntryCreate from './JournalEntryCreate.js';
 import JournalEntry from './JournalEntry.js';
